refactor(docs): extract search document builder
